+++ sdk/storage/azure-storage-blobs/src/blob_sas_builder.cpp
@@ -141,9 +141,9 @@ namespace Azure { namespace Storage { namespace Sas {
         + snapshotVersion + "\n" + CacheControl + "\n" + ContentDisposition + "\n" + ContentEncoding
         + "\n" + ContentLanguage + "\n" + ContentType;
 
-    std::string signature = Base64Encode(Storage::Details::HmacSha256(
+    std::string signature = Azure::Core::Base64Encode(Storage::Details::HmacSha256(
         std::vector<uint8_t>(stringToSign.begin(), stringToSign.end()),
-        Base64Decode(credential.GetAccountKey())));
+        Azure::Core::Base64Decode(credential.GetAccountKey())));
 
     Azure::Core::Http::Url builder;
     builder.AppendQueryParameter(
@@ -241,9 +241,9 @@ namespace Azure { namespace Storage { namespace Sas {
         + CacheControl + "\n" + ContentDisposition + "\n" + ContentEncoding + "\n" + ContentLanguage
         + "\n" + ContentType;
 
-    std::string signature = Base64Encode(Storage::Details::HmacSha256(
+    std::string signature = Azure::Core::Base64Encode(Storage::Details::HmacSha256(
         std::vector<uint8_t>(stringToSign.begin(), stringToSign.end()),
-        Base64Decode(userDelegationKey.Value)));
+        Azure::Core::Base64Decode(userDelegationKey.Value)));
 
     Azure::Core::Http::Url builder;
     builder.AppendQueryParameter(